Move resource parsing into a shared function
